remove console log
